feat: Add support for more visualizations and network graph
