+++ pkg/plugin/datasource.go
@@ -4,17 +4,19 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	// "strings"
 
 	"net/http"
 
 	"github.com/accuknox/kubearmor/pkg/adapters"
 	"github.com/accuknox/kubearmor/pkg/models"
+	"github.com/accuknox/kubearmor/pkg/visualizations"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/httpclient"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
-	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/kubearmor/KubeArmor/KubeArmor/types"
+	// "github.com/grafana/grafana-plugin-sdk-go/data"
+	// "github.com/kubearmor/KubeArmor/KubeArmor/types"
 )
 
 var (
@@ -27,11 +29,6 @@ const (
 	pts0   = "pts0"
 	denied = "Permission denied"
 )
-
-type BackendService interface {
-	GetLogs(ctx context.Context, qm models.QueryModel) ([]types.Log, error)
-	HealthCheck(ctx context.Context) (*backend.CheckHealthResult, error)
-}
 
 // NewDatasource creates a new datasource instance.
 func NewDatasource(ctx context.Context, settings backend.DataSourceInstanceSettings) (instancemgmt.Instance, error) {
@@ -70,7 +67,7 @@ func NewDatasource(ctx context.Context, settings backend.DataSourceInstanceSetti
 
 	datastoreConfig.URL = settings.URL
 	datastoreConfig.Index = PluginSettings.Index
-	backendSVC, err := getBackendService(ctx, Backend, datastoreConfig)
+	backendSVC, err := adapters.GetBackendService(ctx, Backend, datastoreConfig)
 	if err != nil {
 		return nil, err
 	}
@@ -86,7 +83,7 @@ func NewDatasource(ctx context.Context, settings backend.DataSourceInstanceSetti
 // its health and has streaming skills.
 type Datasource struct {
 	settings   backend.DataSourceInstanceSettings
-	BackendSvc BackendService
+	BackendSvc adapters.BackendService
 
 	httpClient *http.Client
 }
@@ -122,7 +119,7 @@ func (d *Datasource) QueryData(ctx context.Context, req *backend.QueryDataReques
 }
 
 func (d *Datasource) query(ctx context.Context, _ backend.PluginContext, q backend.DataQuery) backend.DataResponse {
-	var response backend.DataResponse
+	// var response backend.DataResponse
 
 	// Unmarshal the JSON into our queryModel.
 	var qm models.QueryModel
@@ -133,257 +130,15 @@ func (d *Datasource) query(ctx context.Context, _ backend.PluginContext, q backe
 	if err != nil {
 		ctxLogger.Error("Error while marshalling the query json")
 		return backend.ErrDataResponse(backend.StatusBadRequest, fmt.Sprintf("json unmarshal: %v", err.Error()))
-	} else {
-		ctxLogger.Info("Query json is sucessfully marshalled operation: ")
 	}
+	visualization := visualizations.NewVisualization(ctx, d.BackendSvc, qm)
 
-	Nodegraph, err := getNodeGraph(ctx, d, qm)
-	if err != nil {
-		response.Error = err
-		return response
-	}
-
-	Nodefields := models.GetNodeFields()
-	EdgeFields := models.GetEdgeFields()
-	NetworkFields := models.GetNetworkNodeFields()
-
-	Nodeframe := data.NewFrame("Nodes")
-	if qm.Operation == "Process" {
-
-		Nodeframe.Fields = Nodefields
-	} else {
-		Nodeframe.Fields = NetworkFields
-	}
-
-	EdgeFrame := data.NewFrame("Edges")
-	EdgeFrame.Fields = EdgeFields
-
-	var frameMeta = data.FrameMeta{
-		PreferredVisualization: data.VisTypeNodeGraph,
-	}
-	Nodeframe.SetMeta(&frameMeta)
-	EdgeFrame.SetMeta(&frameMeta)
-
-	for _, node := range Nodegraph.Nodes {
-		if qm.Operation == "Process" {
-
-			Nodeframe.AppendRow(
-				node.ID,
-				node.Title,
-				node.MainStat,
-				node.Color,
-				node.DetailClusterName,
-				node.DetailHostName,
-				node.DetailNamespaceName,
-				node.DetailPodName,
-				node.DetailLabels,
-				node.DetailContainerID,
-				node.DetailContainerName,
-				node.DetailContainerImage,
-				node.DetailParentProcessName,
-				node.DetailProcessName,
-				int64(node.DetailHostPPID),
-				int64(node.DetailHostPID),
-				int64(node.DetailPPID),
-				int64(node.DetailPID),
-				int64(node.DetailUID),
-				node.DetailType,
-				node.DetailSource,
-				node.DetailOperation,
-				node.DetailResource,
-				node.DetailData,
-				node.DetailResult,
-				node.DetailCwd,
-				node.DetailTTY,
-			)
-		} else if qm.Operation == "Network" {
-
-			Nodeframe.AppendRow(
-				node.ID,
-				node.Title,
-				node.MainStat,
-				node.Color,
-				node.DetailPodName,
-				node.DetailNamespaceName,
-			)
-		}
-	}
-
-	for _, edge := range Nodegraph.Edges {
-		EdgeFrame.AppendRow(edge.ID, edge.Source, edge.Target, edge.Mainstat, edge.Count)
-	}
-
-	response.Frames = append(response.Frames, Nodeframe)
-	response.Frames = append(response.Frames, EdgeFrame)
-
-	return response
+	return visualization.GetVisualization()
 }
 
-func getBackendService(ctx context.Context, backendName string, dsc models.DataStoreConfig) (BackendService, error) {
-
-	ctxLogger := log.DefaultLogger.FromContext(ctx)
-	switch backendName {
-
-	case "ELASTICSEARCH":
-		return nil, nil
-
-	case "OPENSEARCH":
-		client, err := adapters.NewOpenSearchClient(dsc)
-		if err != nil {
-			ctxLogger.Error("Cannot create opensearch client %v :", err)
-			return nil, err
-		}
-
-		return client, nil
-	}
-	return nil, nil
-}
-
-func getNodeGraph(ctx context.Context, datasource *Datasource, qm models.QueryModel) (models.NodeGraph, error) {
-	service := datasource.BackendSvc // GetLogs gets the KubeArmor logs from the respective datastore
-	logs, err := service.GetLogs(ctx, qm)
-	if err != nil {
-		return models.NodeGraph{}, err
-	}
-
-	ctxLogger := log.DefaultLogger.FromContext(ctx)
-	ctxLogger.Info(fmt.Sprintf("received logs with len: %d", len(logs)))
-
-	if qm.Operation == "Process" {
-		return getProcessGraph(logs, qm), nil
-	}
-
-	return models.NodeGraph{}, nil
-}
-
-func getProcessGraph(logs []types.Log, MyQuery models.QueryModel) models.NodeGraph {
-
-	colors := []string{"orange", "green", "cyan", "rose"}
-
-	var processLogs []types.Log
-
-	for _, log := range logs {
-
-		if log.TTY == pts0 &&
-			log.Operation == MyQuery.Operation && (MyQuery.NamespaceQuery == "All" || log.NamespaceName == MyQuery.NamespaceQuery) &&
-			(MyQuery.LabelQuery == "All" || log.Labels == MyQuery.LabelQuery) {
-			processLogs = append(processLogs, log)
-		}
-
-	}
-
-	/* Nodes */
-
-	var ProcessNodes []models.NodeFields
-
-	var processEdges []models.EdgeFields
-
-	nodeMap := make(map[string]interface{})
-
-	for _, log := range processLogs {
-		isBlocked := log.Result == denied
-
-		if log.PPID == 0 {
-			colorIndex := random(0, len(colors)-1)
-			cnode := models.NodeFields{
-				ID:                  log.ContainerName + log.NamespaceName,
-				Title:               log.ContainerName,
-				Color:               colors[colorIndex],
-				ChildNode:           fmt.Sprintf("%d%s", log.HostPID, log.PodName),
-				DetailContainerName: log.ContainerName,
-				DetailNamespaceName: log.NamespaceName,
-			}
-
-			ProcessNodes = append(ProcessNodes, cnode)
-
-			edge := models.EdgeFields{
-				ID:     fmt.Sprintf("%s%s%s%s", cnode.ID, cnode.ChildNode, cnode.DetailNamespaceName, cnode.DetailContainerName),
-				Source: fmt.Sprintf("%s", cnode.ID),
-				Target: fmt.Sprintf("%s", cnode.ChildNode),
-
-				Mainstat: fmt.Sprintf("%s", "ContainerEdge"),
-				Count:    "None",
-			}
-			nodeMap[cnode.ID] = ""
-
-			processEdges = append(processEdges, edge)
-
-		} else {
-
-			edge := models.EdgeFields{
-				ID:       fmt.Sprintf("%s%d%d", fmt.Sprintf("%d%s%s", log.HostPID, log.ContainerName, log.NamespaceName), log.PPID, log.HostPID),
-				Source:   fmt.Sprintf("%d%s", log.HostPPID, log.PodName),
-				Target:   fmt.Sprintf("%d%s", log.HostPID, log.PodName),
-				Mainstat: fmt.Sprintf("%s", log.Data),
-
-				Count: "None",
-			}
-			processEdges = append(processEdges, edge)
-		}
-		nodeId := fmt.Sprintf("%d%s", log.HostPID, log.PodName)
-		nodeMap[nodeId] = ""
-		node := models.NodeFields{
-			ID:       nodeId,
-			Title:    log.ProcessName,
-			MainStat: log.Source,
-			Color:    "white",
-			// DetailTimestamp:         log.Timestamp,
-			// NodeRadius:              "5",
-			DetailClusterName:       log.ClusterName,
-			DetailHostName:          log.HostName,
-			DetailNamespaceName:     log.NamespaceName,
-			DetailPodName:           log.ContainerName, // Using ContainerName as PodName for demonstration
-			DetailLabels:            log.Labels,
-			DetailContainerID:       log.ContainerID,
-			DetailContainerName:     log.ContainerName,
-			DetailContainerImage:    log.ContainerImage,
-			DetailParentProcessName: log.ParentProcessName,
-			DetailProcessName:       log.ProcessName,
-			DetailHostPPID:          int64(log.HostPPID),
-			DetailHostPID:           int64(log.HostPID),
-			DetailPPID:              int64(log.PPID),
-			DetailPID:               int64(log.PID),
-			DetailUID:               int64(log.UID),
-			DetailType:              log.Type,
-			DetailSource:            log.Source,
-			DetailOperation:         log.Operation,
-			DetailResource:          log.Resource,
-			DetailData:              log.Data,
-			DetailResult:            log.Result,
-			DetailCwd:               log.Cwd,
-			DetailTTY:               log.TTY,
-		}
-
-		if isBlocked {
-			node.Color = "red"
-		}
-
-		ProcessNodes = append(ProcessNodes, node)
-
-	}
-
-	NewprocessEdges := []models.EdgeFields{}
-
-	for _, edge := range processEdges {
-		if _, ok := nodeMap[edge.Source]; ok {
-			if _, ok := nodeMap[edge.Target]; ok {
-				NewprocessEdges = append(NewprocessEdges, edge)
-			}
-		}
-	}
-
-	var nodeGraph = models.NodeGraph{
-		Nodes: ProcessNodes,
-		Edges: NewprocessEdges,
-	}
-
-	return nodeGraph
-}
-
-func getNGraph(logs []types.Log) models.NodeGraph {
-	return models.NodeGraph{}
-}
-
+//	func getNGraph(logs []types.Log) models.NodeGraph {
+//		return models.NodeGraph{}
+//	}
 func (d *Datasource) CheckHealth(ctx context.Context, req *backend.CheckHealthRequest) (*backend.CheckHealthResult, error) {
 	backendService := d.BackendSvc
 	return backendService.HealthCheck(ctx)
