+++ pkg/utils/utils.go
@@ -1,4 +1,4 @@
-package plugin
+package utils
 
 import (
 	"fmt"
